Removed InternalAU from STK physical models
