+++ Sources/AudioKit/Nodes/Generators/Physical Models/Clarinet.swift
@@ -13,37 +13,10 @@ public class Clarinet: Node, AudioUnitContainer, Toggleable {
     public static let ComponentDescription = AudioComponentDescription(instrument: "clar")
 
     /// Internal type of audio unit for this node
-    public typealias AudioUnitType = InternalAU
+    public typealias AudioUnitType = STKAudioUnit
 
     /// Internal audio unit
     public private(set) var internalAU: AudioUnitType?
-
-
-    /// INternal audio unti for clarinet
-    public class InternalAU: AudioUnitBase {
-
-        /// Create the clarinet DSP
-        /// - Returns: DSP Reference
-        public override func createDSP() -> DSPRef {
-            return akCreateDSP("ClarinetDSP")
-        }
-
-        /// Trigger a clarinet note
-        /// - Parameters:
-        ///   - note: MIDI Note Number
-        ///   - velocity: MIDI Velocity
-        public func trigger(note: MIDINoteNumber, velocity: MIDIVelocity) {
-
-            if let midiBlock = scheduleMIDIEventBlock {
-                let event = MIDIEvent(noteOn: note, velocity: velocity, channel: 0)
-                event.data.withUnsafeBufferPointer { ptr in
-                    guard let ptr = ptr.baseAddress else { return }
-                    midiBlock(AUEventSampleTimeImmediate, 0, event.data.count, ptr)
-                }
-            }
-
-        }
-    }
 
     // MARK: - Initialization
