+++ Sources/AudioKit/Nodes/Generators/Physical Models/MandolinString.swift
@@ -13,35 +13,10 @@ public class MandolinString: Node, AudioUnitContainer, Toggleable {
     public static let ComponentDescription = AudioComponentDescription(instrument: "mand")
 
     /// Internal type of audio unit for this node
-    public typealias AudioUnitType = InternalAU
+    public typealias AudioUnitType = STKAudioUnit
 
     /// Internal audio unit
     public private(set) var internalAU: AudioUnitType?
-
-    /// Internal audio unit for mandolin string
-    public class InternalAU: AudioUnitBase {
-
-        /// Create mandolin string DSP
-        /// - Returns: DSP reference
-        public override func createDSP() -> DSPRef {
-            return akCreateDSP("MandolinStringDSP")
-        }
-
-        /// Trigger a mandolin string note
-        /// - Parameters:
-        ///   - note: MIDI Note Number
-        ///   - velocity: MIDI Velocity
-        public func trigger(note: MIDINoteNumber, velocity: MIDIVelocity) {
-            if let midiBlock = scheduleMIDIEventBlock {
-                let event = MIDIEvent(noteOn: note, velocity: velocity, channel: 0)
-                event.data.withUnsafeBufferPointer { ptr in
-                    guard let ptr = ptr.baseAddress else { return }
-                    midiBlock(AUEventSampleTimeImmediate, 0, event.data.count, ptr)
-                }
-            }
-
-        }
-    }
 
     // MARK: - Initialization
