+++ Sources/CAudioKit/Nodes/Generators/Physical Models/ClarinetDSP.mm
@@ -31,4 +31,4 @@ public:
 
 };
 
-AK_REGISTER_DSP(ClarinetDSP);
+AK_REGISTER_DSP2(ClarinetDSP, "clar");
