+++ Sources/CAudioKit/Nodes/Generators/Physical Models/FluteDSP.mm
@@ -31,4 +31,4 @@ public:
 
 };
 
-AK_REGISTER_DSP(FluteDSP);
+AK_REGISTER_DSP2(FluteDSP, "flut");
