+++ Sources/CAudioKit/Nodes/Generators/Physical Models/MandolinStringDSP.mm
@@ -61,4 +61,4 @@ public:
 
 };
 
-AK_REGISTER_DSP(MandolinStringDSP);
+AK_REGISTER_DSP2(MandolinStringDSP, "mand");
