+++ Sources/CAudioKit/Nodes/Generators/Physical Models/RhodesPianoKeyDSP.mm
@@ -50,4 +50,4 @@ public:
 
 };
 
-AK_REGISTER_DSP(RhodesPianoKeyDSP);
+AK_REGISTER_DSP2(RhodesPianoKeyDSP, "rhds");
